Buffer initialisation and log config is now in shared folder
