+++ src/pages/App/App.test.tsx
@@ -2,63 +2,17 @@ import { App } from "@pages";
 import { render } from "@testing-library/react";
 import renderer from "react-test-renderer";
 import "@testing-library/jest-dom";
-import log from "loglevel";
 
 describe("Test the demo component", () => {
 
-  test("renders HASKI basic label", () => {
-    const landingPage = render(<App />);
-    expect(landingPage.getByText("Hello with ID:")).toBeInTheDocument();
-  });
-
-  test("renders correctly", () => {
-    const tree = renderer.create(<App />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  test("log messages are added to localStorage", () => {
-    render(<App />);
-    log.setLevel("trace");
-    log.trace("This is a trace");
-    log.warn("This is a warning");
-    log.error("This is an error");
-    const currentDate = new Date().toUTCString();
-
-    expect(localStorage.getItem('ringBufferContent')).not.toBeNull();
-    expect(localStorage.getItem('ringBufferContent')).toContain(currentDate);
-    expect(localStorage.getItem('ringBufferContent')).toContain("This is a warning");
-    expect(localStorage.getItem('ringBufferContent')).toContain("This is an error");
-    expect(localStorage.getItem('ringBufferContent')).toContain("This is a trace");
-  });
-
-
-});
-
-describe("localStorage already set", () => {
-
-    beforeEach(() => {
-        Object.defineProperty(window, "localStorage", {
-            value: {
-                getItem: jest.fn(() => '{"buffer": ["mockedValue"]}'),
-                setItem: localStorage.setItem.bind(localStorage),
-                removeItem: jest.fn(),
-            },
-            writable: true,
-        });
-        localStorage.setItem("ringBufferContent", '{"buffer": ["mockedValue"]}');
+    test("renders HASKI basic label", () => {
+        const landingPage = render(<App/>);
+        expect(landingPage.getByText("Hello with ID:")).toBeInTheDocument();
     });
 
-    test("should return a value from GlobalRingBuffer", () => {
-
-        render(<App />);
-        log.setLevel("trace");
-        log.trace("This is a trace");
-        log.warn("This is a warning");
-        log.error("This is an error");
-
-        expect(localStorage.getItem('ringBufferContent')).not.toBeNull();
-        expect(localStorage.getItem('ringBufferContent')).toContain("{\"buffer\": [\"mockedValue\"]}");
-
+    test("renders correctly", () => {
+        const tree = renderer.create(<App/>).toJSON();
+        expect(tree).toMatchSnapshot();
     });
 
 });
